+++ GUI/EXCEL/excelNhanVien.cs.cs
@@ -162,7 +162,7 @@ namespace GUI.EXCEL
                                 HoTen = hoTen,
                                 SoDienThoai = soDienThoai,
                                 Email = email,
-                                Luong = decimal.TryParse(ws.Cells[row, 5].Text, out decimal luong) ? luong : 0,
+                                Luong = float.TryParse(ws.Cells[row, 5].Text, out float luong) ? luong : 0,
                                 NgayTao = DateTime.Now
                             };
